requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.2
django==5.2.1
django-filter==25.1
django-shortuuidfield==0.1.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
markdown==3.8
pillow==11.2.1
pip==24.0
pycparser==2.22
pyjwt==2.9.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
uv==0.7.5
